fix: rename function,update comment
